Your Answer
I began to go through your code sir and spent the last few weeks understanding java, but when I went to do it I looked at your solution to understand it. It was completed and I was confused as I had understood it and could not improve it. I went to add notes but notes were already there.